requirements.txt
# Bot – zamonaviy versiyalar (Render deploy uchun)
aiogram>=3.13.0,<4
yt-dlp>=2024.12.0
shazamio>=0.7.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
mutagen>=1.47.0
aiosqlite>=0.19.0
Pillow>=10.0.0
pydub>=0.25.0

# FFmpeg (Dockerda tizimda; lokalda ixtiyoriy)
imageio-ffmpeg>=0.5.0
